Reestructure autoconfigure module
Fixes gh-1
