--- autoconfigure/http/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-zipkin2.autoconfigure.storage.forwarder.ZipkinHttpForwarderStorageAutoConfiguration
