--- autoconfigure/kafka/pom.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright 2019 Jorge Esteban Quilcate Otoya
-
-    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
-    in compliance with the License. You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software distributed under the License
-    is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-    or implied. See the License for the specific language governing permissions and limitations under
-    the License.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>io.github.jeqo.zipkin</groupId>
-    <artifactId>zipkin-forwarder-autoconfigure</artifactId>
-    <version>0.1.3-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>zipkin-forwarder-kafka-autoconfigure</artifactId>
-
-  <properties>
-    <main.basedir>${project.basedir}/../..</main.basedir>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>io.github.jeqo.zipkin</groupId>
-      <artifactId>zipkin-forwarder-kafka</artifactId>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-maven-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-</project>
