+++ autoconfigure/src/main/resources/zipkin-server-http-forwarder.yml
