+++ autoconfigure/src/main/resources/zipkin-server-kafka-forwarder.yml
