+++ autoconfigure/src/test/java/zipkin2/autoconfure/storage/reporter/HttpAccessReporter.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2019 Jorge Esteban Quilcate Otoya
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package zipkin2.autoconfure.storage.reporter;
+
+import org.springframework.boot.autoconfigure.context.PropertyPlaceholderAutoConfiguration;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import zipkin2.autoconfigure.storage.forwarder.http.ZipkinHttpForwarderStorageAutoConfiguration;
+
+public class HttpAccessReporter {
+
+    public static void registerReporter(AnnotationConfigApplicationContext context) {
+        context.register(PropertyPlaceholderAutoConfiguration.class, ZipkinHttpForwarderStorageAutoConfiguration.class);
+    }
+}
